+++ packages/video_player_avplay/lib/src/closed_caption_file.dart
@@ -2,8 +2,12 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:flutter/foundation.dart' show immutable, objectRuntimeType;
+import 'dart:typed_data' show Uint8List;
+import 'package:flutter/foundation.dart'
+    show immutable, listEquals, objectRuntimeType;
+import 'package:flutter/material.dart';
 
+import '../video_player_platform_interface.dart';
 import 'sub_rip.dart';
 import 'web_vtt.dart';
 
@@ -25,7 +29,7 @@ abstract class ClosedCaptionFile {
   /// The full list of captions from a given file.
   ///
   /// The [captions] will be in the order that they appear in the given file.
-  List<Caption> get captions;
+  List<TextCaption> get textCaptions;
 }
 
 /// A representation of a single caption.
@@ -33,17 +37,14 @@ abstract class ClosedCaptionFile {
 /// A typical closed captioning file will include several [Caption]s, each
 /// linked to a start and end time.
 @immutable
-class Caption {
+sealed class Caption {
   /// Creates a new [Caption] object.
   ///
-  /// This is not recommended for direct use unless you are writing a parser for
-  /// a new closed captioning file type.
+  /// This constructor is for internal use by subclasses.
   const Caption({
     required this.number,
     required this.start,
     required this.end,
-    required this.text,
-    this.subtitleAttributes,
   });
 
   /// The number that this caption was assigned.
@@ -54,17 +55,59 @@ class Caption {
 
   /// When in the given video should this [Caption] be dismissed.
   final Duration end;
+}
+
+/// A representation of a text-based caption.
+@immutable
+class TextCaption extends Caption {
+  /// Creates a new [TextCaption] object.
+  const TextCaption(
+      {required super.number,
+      required super.start,
+      required super.end,
+      required this.text,
+      this.subtitleAttributes,
+      this.textStyle,
+      this.textOriginAndExtent,
+      this.textAlign,
+      this.windowBgColor,
+      this.fontSize,
+      this.fontForeground});
 
   /// The actual text that should appear on screen to be read between [start]
   /// and [end].
   final String text;
 
-  /// The subtitile attributes associated with this caption.
+  /// The subtitle attributes associated with this caption.
   final List<SubtitleAttribute?>? subtitleAttributes;
 
-  /// A no caption object. This is a caption with [start] and [end] durations of zero,
+  /// Specifies how the text in the closed caption should look.
+  final TextStyle? textStyle;
+
+  /// The origin and extent of the text in the closed caption.
+  final TextOriginAndExtent? textOriginAndExtent;
+
+  /// The alignment of the text in the closed caption.
+  final AlignmentGeometry? textAlign;
+
+  /// The background color of the window where the text appears.
+  final Color? windowBgColor;
+
+  /// The font size of the text in the closed caption.
+  final double? fontSize;
+
+  /// The outline of the text in the closed caption.
+  final Paint? fontForeground;
+
+  /// The length of the hexadecimal string representing an RGB color.
+  static const int lengthRGB = 6;
+
+  /// The length of the hexadecimal string representing an ARGB color.
+  static const int lengthARGB = 8;
+
+  /// A no text caption object. This is a caption with [start] and [end] durations of zero,
   /// and an empty [text] string.
-  static const Caption none = Caption(
+  static const TextCaption none = TextCaption(
     number: 0,
     start: Duration.zero,
     end: Duration.zero,
@@ -72,29 +115,345 @@ class Caption {
     subtitleAttributes: <SubtitleAttribute?>[],
   );
 
+  static Color _intToColor(int colorValue) {
+    if (colorValue <= 0) {
+      return const Color(0x00000000);
+    }
+    String hexValue = colorValue.toRadixString(16);
+    if (hexValue.length < lengthRGB) {
+      hexValue = hexValue.padLeft(lengthRGB, '0');
+    }
+    if (hexValue.length == lengthRGB) {
+      hexValue = 'FF$hexValue';
+    }
+    if (hexValue.length == lengthARGB) {
+      return Color(int.parse('0x$hexValue'));
+    }
+    return const Color(0x00000000);
+  }
+
+  static FontWeight _intToFontWeight(int weightValue) {
+    return weightValue == 0 ? FontWeight.normal : FontWeight.bold;
+  }
+
+  static FontStyle _intToFontStyle(int styleValue) {
+    return styleValue == 0 ? FontStyle.normal : FontStyle.italic;
+  }
+
+  static AlignmentGeometry _intToTextVerticalAlign(int alignValue) {
+    return switch (alignValue) {
+      0 => Alignment.topCenter,
+      2 => Alignment.bottomCenter,
+      _ => Alignment.center,
+    };
+  }
+
+  static AlignmentGeometry _intToTextHorizontalAlign(int alignValue) {
+    return switch (alignValue) {
+      0 => Alignment.centerLeft,
+      2 => Alignment.centerRight,
+      _ => Alignment.center,
+    };
+  }
+
+  /// Process subtitle attributes and return the text alignment and text style.
+  static (
+    TextOriginAndExtent?,
+    TextStyle?,
+    AlignmentGeometry,
+    Color,
+    double,
+    Paint
+  ) processSubtitleAttributes(List<SubtitleAttribute> subtitleAttributes) {
+    TextOriginAndExtent actualTextOriginAndExtent = TextOriginAndExtent.none;
+    TextStyle actualTextStyle =
+        const TextStyle(height: 1.0, fontSize: 1 / 15.0);
+    AlignmentGeometry actualTextAlign = Alignment.center;
+    Color actualWindowBgColor =
+        const Color(0x00000000); // default transparent color.
+    double actualFontSize = 1 / 15.0;
+    Paint foreground = Paint()..style = PaintingStyle.stroke;
+
+    if (subtitleAttributes.isEmpty) {
+      return (
+        actualTextOriginAndExtent,
+        actualTextStyle,
+        actualTextAlign,
+        actualWindowBgColor,
+        actualFontSize,
+        foreground
+      );
+    }
+
+    for (final SubtitleAttribute attr in subtitleAttributes) {
+      switch (attr.attrType) {
+        // For text origin and extent.
+        case SubtitleAttrType.subAttrRegionXPos:
+          final double xPos = attr.attrValue as double;
+          if (xPos > 0) {
+            actualTextOriginAndExtent =
+                actualTextOriginAndExtent.addValue(originX: xPos);
+          }
+        case SubtitleAttrType.subAttrRegionYPos:
+          final double yPos = attr.attrValue as double;
+          if (yPos > 0) {
+            actualTextOriginAndExtent =
+                actualTextOriginAndExtent.addValue(originY: yPos);
+          }
+        case SubtitleAttrType.subAttrRegionWidth:
+          final double width = attr.attrValue as double;
+          if (width > 0) {
+            actualTextOriginAndExtent =
+                actualTextOriginAndExtent.addValue(extentWidth: width);
+          }
+        case SubtitleAttrType.subAttrRegionHeight:
+          final double height = attr.attrValue as double;
+          if (height > 0) {
+            actualTextOriginAndExtent =
+                actualTextOriginAndExtent.addValue(extentHeight: height);
+          }
+        // For text style.
+        case SubtitleAttrType.subAttrFontFamily:
+          actualTextStyle =
+              actualTextStyle.copyWith(fontFamily: attr.attrValue as String);
+        case SubtitleAttrType.subAttrFontSize:
+          final double fontSize = attr.attrValue as double;
+          if (fontSize > 0) {
+            actualFontSize = fontSize;
+          }
+        case SubtitleAttrType.subAttrFontWeight:
+          actualTextStyle = actualTextStyle.copyWith(
+              fontWeight: _intToFontWeight(attr.attrValue as int));
+        case SubtitleAttrType.subAttrFontStyle:
+          actualTextStyle = actualTextStyle.copyWith(
+              fontStyle: _intToFontStyle(attr.attrValue as int));
+        case SubtitleAttrType.subAttrFontColor:
+          if (actualTextStyle.foreground == null) {
+            actualTextStyle = actualTextStyle.copyWith(
+                color: _intToColor(attr.attrValue as int));
+          }
+        case SubtitleAttrType.subAttrFontBgColor:
+          if (actualTextStyle.background == null) {
+            actualTextStyle = actualTextStyle.copyWith(
+                backgroundColor: _intToColor(attr.attrValue as int));
+          }
+        case SubtitleAttrType.subAttrFontOpacity:
+          final double fontOpacity = attr.attrValue as double;
+          if (fontOpacity >= 0.0 && fontOpacity <= 1.0) {
+            if (actualTextStyle.color == null) {
+              actualTextStyle = actualTextStyle.copyWith(color: Colors.white);
+            }
+            actualTextStyle = actualTextStyle.copyWith(
+                color: actualTextStyle.color!.withValues(alpha: fontOpacity));
+          }
+        case SubtitleAttrType.subAttrFontBgOpacity:
+          final double fontBgOpacity = attr.attrValue as double;
+          if (fontBgOpacity >= 0.0 &&
+              fontBgOpacity <= 1.0 &&
+              actualTextStyle.backgroundColor != null) {
+            actualTextStyle = actualTextStyle.copyWith(
+                backgroundColor: actualTextStyle.backgroundColor!
+                    .withValues(alpha: fontBgOpacity));
+          }
+        // For text vertical and horizontal align.
+        case SubtitleAttrType.subAttrFontVerticalAlign:
+          actualTextAlign = actualTextAlign
+              .add(_intToTextVerticalAlign(attr.attrValue as int));
+        case SubtitleAttrType.subAttrFontHorizontalAlign:
+          actualTextAlign = actualTextAlign
+              .add(_intToTextHorizontalAlign(attr.attrValue as int));
+        // For text background window.
+        case SubtitleAttrType.subAttrWindowBgColor:
+          actualWindowBgColor = _intToColor(attr.attrValue as int);
+        case SubtitleAttrType.subAttrWindowOpacity:
+          final double windowOpacity = attr.attrValue as double;
+          if (windowOpacity >= 0.0 && windowOpacity <= 1.0) {
+            actualWindowBgColor =
+                actualWindowBgColor.withValues(alpha: windowOpacity);
+          }
+        // For text outline.
+        case SubtitleAttrType.subAttrFontTextOutlineColor:
+          foreground = foreground..color = _intToColor(attr.attrValue as int);
+        case SubtitleAttrType.subAttrFontTextOutlineThickness:
+          foreground = foreground..strokeWidth = (attr.attrValue as int) / 1.0;
+        case SubtitleAttrType.subAttrFontTextOutlineBlurRadius:
+        case SubtitleAttrType.subAttrWindowXPadding:
+        case SubtitleAttrType.subAttrWindowYPadding:
+        case SubtitleAttrType.subAttrWindowLeftMargin:
+        case SubtitleAttrType.subAttrWindowRightMargin:
+        case SubtitleAttrType.subAttrWindowTopMargin:
+        case SubtitleAttrType.subAttrWindowBottomMargin:
+        case SubtitleAttrType.subAttrWindowShowBg:
+        case SubtitleAttrType.subAttrRawSubtitle:
+        case SubtitleAttrType.subAttrWebvttCueLine:
+        case SubtitleAttrType.subAttrWebvttCueLineNum:
+        case SubtitleAttrType.subAttrWebvttCueLineAlign:
+        case SubtitleAttrType.subAttrWebvttCueAlign:
+        case SubtitleAttrType.subAttrWebvttCueSize:
+        case SubtitleAttrType.subAttrWebvttCuePosition:
+        case SubtitleAttrType.subAttrWebvttCuePositionAlign:
+        case SubtitleAttrType.subAttrWebvttCueVertical:
+        case SubtitleAttrType.subAttrTimestamp:
+        case SubtitleAttrType.subAttrExtsubIndex:
+        case SubtitleAttrType.subAttrTypeNone:
+          break;
+      }
+    }
+    return (
+      actualTextOriginAndExtent == TextOriginAndExtent.none
+          ? null
+          : actualTextOriginAndExtent,
+      actualTextStyle == const TextStyle(height: 1.0, fontSize: 1 / 15.0)
+          ? null
+          : actualTextStyle,
+      actualTextAlign,
+      actualWindowBgColor,
+      actualFontSize,
+      foreground
+    );
+  }
+
   @override
   String toString() {
-    return '${objectRuntimeType(this, 'Caption')}('
+    return '${objectRuntimeType(this, 'TextCaption')}('
         'number: $number, '
         'start: $start, '
         'end: $end, '
         'text: $text, '
-        'subtitleAttributes: $subtitleAttributes)';
+        'subtitleAttributes: $subtitleAttributes, '
+        'textStyle: $textStyle)';
   }
 
   @override
   bool operator ==(Object other) =>
       identical(this, other) ||
-      other is Caption &&
+      other is TextCaption &&
           runtimeType == other.runtimeType &&
           number == other.number &&
           start == other.start &&
           end == other.end &&
           text == other.text &&
-          subtitleAttributes == other.subtitleAttributes;
+          listEquals(subtitleAttributes, other.subtitleAttributes) &&
+          textStyle == other.textStyle;
 
   @override
-  int get hashCode => Object.hash(number, start, end, text, subtitleAttributes);
+  int get hashCode =>
+      Object.hash(number, start, end, text, subtitleAttributes, textStyle);
+}
+
+/// A representation of a picture-based caption.
+@immutable
+class PictureCaption extends Caption {
+  /// Creates a new [PictureCaption] object.
+  const PictureCaption({
+    required super.number,
+    required super.start,
+    required super.end,
+    this.picture,
+    this.pictureWidth,
+    this.pictureHeight,
+  });
+
+  /// The image data for the caption, typically in a format like PNG or JPEG.
+  final Uint8List? picture;
+
+  /// Specifies the picture's width.
+  final double? pictureWidth;
+
+  /// Specifies the picture's height.
+  final double? pictureHeight;
+
+  /// A no picture caption object. This is a caption with [start] and [end] durations of zero,
+  /// and an empty [picture].
+  static const PictureCaption none = PictureCaption(
+    number: 0,
+    start: Duration.zero,
+    end: Duration.zero,
+    pictureWidth: 0.0,
+    pictureHeight: 0.0,
+  );
+
+  @override
+  String toString() {
+    return '${objectRuntimeType(this, 'PictureCaption')}('
+        'number: $number, '
+        'start: $start, '
+        'end: $end, '
+        'picture: ${picture == null ? 'null' : '${picture?.length} bytes'}, '
+        'pictureWidth: $pictureWidth, '
+        'pictureHeight: $pictureHeight)';
+  }
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is PictureCaption &&
+          runtimeType == other.runtimeType &&
+          number == other.number &&
+          start == other.start &&
+          end == other.end &&
+          picture == other.picture &&
+          pictureWidth == other.pictureWidth &&
+          pictureHeight == other.pictureHeight;
+
+  @override
+  int get hashCode =>
+      Object.hash(number, start, end, picture, pictureWidth, pictureHeight);
+}
+
+/// This attribute defines the upper-left corner of a rectangular region using [originX] and [originY] coordinate values
+/// and the width and height of the rectangular region using [extentWidth] and [extentHeight],
+/// which are typically specified as percentages of the video's width and height.
+@immutable
+class TextOriginAndExtent {
+  /// Creates a new [TextOriginAndExtent] object.
+  const TextOriginAndExtent(
+      {required this.originX,
+      required this.originY,
+      this.extentWidth,
+      this.extentHeight});
+
+  /// The x-coordinate of the upper-left corner of the rectangular region.
+  final double originX;
+
+  /// The y-coordinate of the upper-left corner of the rectangular region.
+  final double originY;
+
+  /// The width of the rectangular region.
+  final double? extentWidth;
+
+  /// The height of the rectangular region.
+  final double? extentHeight;
+
+  /// Adds values to the current [TextOriginAndExtent] object.
+  TextOriginAndExtent addValue(
+      {double? originX,
+      double? originY,
+      double? extentWidth,
+      double? extentHeight}) {
+    return TextOriginAndExtent(
+        originX: originX ?? this.originX,
+        originY: originY ?? this.originY,
+        extentWidth: extentWidth ?? this.extentWidth,
+        extentHeight: extentHeight ?? this.extentHeight);
+  }
+
+  /// A no text origin and extent object. This is a caption with [originX], [originY], [extentWidth], and [extentHeight] of zero.
+  static const TextOriginAndExtent none = TextOriginAndExtent(
+      originX: 0.0, originY: 0.0, extentWidth: 0.0, extentHeight: 0.0);
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is TextOriginAndExtent &&
+          runtimeType == other.runtimeType &&
+          originX == other.originX &&
+          originY == other.originY &&
+          extentWidth == other.extentWidth &&
+          extentHeight == other.extentHeight;
+
+  @override
+  int get hashCode => Object.hash(originX, originY, extentWidth, extentHeight);
 }
 
 /// The different types of subtitle attributes that can be set on the player.
@@ -208,7 +567,7 @@ enum SubtitleAttrType {
   subAttrTimestamp(SubtitleAttrValueType.int),
 
   /// File index of external subtitle
-  subAttrExtsubInde(SubtitleAttrValueType.none),
+  subAttrExtsubIndex(SubtitleAttrValueType.none),
 
   /// Default type
   subAttrTypeNone(SubtitleAttrValueType.none);
@@ -281,32 +640,34 @@ class SubtitleAttribute {
         eventSubtitleAttrList!.cast<Map<Object?, Object?>?>();
 
     for (final Map<Object?, Object?>? attr in subtitleAttrList) {
-      final int attrTypeNum = attr!['attrType']! as int;
-      final int startTime = attr['startTime']! as int;
-      final int stopTime = attr['stopTime']! as int;
+      if (attr != null && attr['attrType'] != null) {
+        final int attrTypeNum = attr['attrType']! as int;
+        final int startTime = attr['startTime']! as int;
+        final int stopTime = attr['stopTime']! as int;
 
-      Object attrValue;
-      if (SubtitleAttrType.getValueType(attrTypeNum) ==
-          SubtitleAttrValueType.double) {
-        attrValue = attr['attrValue']! as double;
-      } else if (SubtitleAttrType.getValueType(attrTypeNum) ==
-          SubtitleAttrValueType.int) {
-        attrValue = attr['attrValue']! as int;
-      } else if (SubtitleAttrType.getValueType(attrTypeNum) ==
-          SubtitleAttrValueType.string) {
-        attrValue = attr['attrValue']! as String;
-      } else {
-        attrValue = 'failed';
+        Object attrValue;
+        if (SubtitleAttrType.getValueType(attrTypeNum) ==
+            SubtitleAttrValueType.double) {
+          attrValue = attr['attrValue']! as double;
+        } else if (SubtitleAttrType.getValueType(attrTypeNum) ==
+            SubtitleAttrValueType.int) {
+          attrValue = attr['attrValue']! as int;
+        } else if (SubtitleAttrType.getValueType(attrTypeNum) ==
+            SubtitleAttrValueType.string) {
+          attrValue = attr['attrValue']! as String;
+        } else {
+          attrValue = 'failed';
+        }
+
+        subtitleAttributes.add(
+          SubtitleAttribute(
+            attrType: SubtitleAttrType.values[attrTypeNum],
+            startTime: startTime,
+            stopTime: stopTime,
+            attrValue: attrValue,
+          ),
+        );
       }
-
-      subtitleAttributes.add(
-        SubtitleAttribute(
-          attrType: SubtitleAttrType.values[attrTypeNum],
-          startTime: startTime,
-          stopTime: stopTime,
-          attrValue: attrValue,
-        ),
-      );
     }
     return subtitleAttributes;
   }
@@ -332,4 +693,120 @@ class SubtitleAttribute {
 
   @override
   int get hashCode => Object.hash(attrType, startTime, stopTime, attrValue);
+}
+
+/// A representation of the caption which includes text or picture.
+@immutable
+class Captions {
+  /// Creates a new [Captions] object.
+  const Captions({this.textCaptions, this.pictureCaption});
+
+  /// The [TextCaption] that should be displayed based on the current [position].
+  ///
+  ///  If there is no text caption for the current
+  /// [position], this will be a [TextCaption.none] object.
+  final List<TextCaption>? textCaptions;
+
+  /// The [PictureCaption] that should be displayed based on the current [position].
+  ///
+  ///  If there is no text caption for the current
+  /// [position], this will be a [PictureCaption.none] object.
+  final PictureCaption? pictureCaption;
+
+  /// Parses a subtitle from a [VideoEvent.subtitlesInfo] into a [Captions] object.
+  static Captions? parseSubtitle(
+      Duration position, SubtitlesInfo subtitlesInfo) {
+    final Duration textDuration = subtitlesInfo.textDuration == 0
+        ? Duration.zero
+        : Duration(milliseconds: subtitlesInfo.textDuration!);
+    if (subtitlesInfo.pictureInfo?.isNotEmpty ?? false) {
+      final PictureCaption pictureCaption = PictureCaption(
+        number: 0,
+        start: position,
+        end: position + textDuration,
+        picture: subtitlesInfo.pictureInfo!['picture'] as Uint8List?,
+        pictureWidth: subtitlesInfo.pictureInfo!['pictureWidth'] as double?,
+        pictureHeight: subtitlesInfo.pictureInfo!['pictureHeight'] as double?,
+      );
+      return Captions(
+          textCaptions: const <TextCaption>[TextCaption.none],
+          pictureCaption: pictureCaption);
+    } else {
+      final int textLines = (subtitlesInfo.textsInfo == null ||
+              subtitlesInfo.textsInfo![0] == null)
+          ? 0
+          : subtitlesInfo.textsInfo!.length;
+
+      if (textLines > 0) {
+        final List<TextCaption> textCaptions = <TextCaption>[];
+        for (int i = 0; i < textLines; i++) {
+          final Map<Object?, Object?> textInfo =
+              subtitlesInfo.textsInfo![i] as Map<Object?, Object?>;
+          final String? text = textInfo['text'] as String?;
+          final List<dynamic>? subtitleAttrList =
+              textInfo['attributes'] as List<dynamic>?;
+
+          final List<SubtitleAttribute> subtitleAttributes =
+              SubtitleAttribute.fromEventSubtitleAttrList(subtitleAttrList);
+
+          final (
+            TextOriginAndExtent?,
+            TextStyle?,
+            AlignmentGeometry,
+            Color,
+            double,
+            Paint
+          ) subtitleAttr =
+              TextCaption.processSubtitleAttributes(subtitleAttributes);
+          final TextOriginAndExtent? actualTextOriginAndExtent =
+              subtitleAttr.$1;
+          final TextStyle? actualTextStyle = subtitleAttr.$2;
+          final AlignmentGeometry actualTextAlign = subtitleAttr.$3;
+          final Color actualWindowBgColor = subtitleAttr.$4;
+          final double actualFontSize = subtitleAttr.$5;
+          final Paint actualForeground = subtitleAttr.$6;
+
+          final TextCaption textCaption = TextCaption(
+              number: 0,
+              start: position,
+              end: position + textDuration,
+              text: text ?? '',
+              subtitleAttributes: subtitleAttributes,
+              textStyle: actualTextStyle,
+              textOriginAndExtent: actualTextOriginAndExtent,
+              textAlign: actualTextAlign,
+              windowBgColor: actualWindowBgColor,
+              fontSize: actualFontSize,
+              fontForeground: actualForeground);
+
+          textCaptions.add(textCaption);
+        }
+        return Captions(
+            textCaptions: textCaptions, pictureCaption: PictureCaption.none);
+      }
+    }
+    return null;
+  }
+
+  /// A no captions object. This is a caption with a list of [TextCaption.none] and a [PictureCaption.none].
+  static const Captions none = Captions(
+      textCaptions: <TextCaption>[TextCaption.none],
+      pictureCaption: PictureCaption.none);
+
+  @override
+  String toString() {
+    return '${objectRuntimeType(this, 'Captions')}('
+        'textCaptions: $textCaptions, '
+        'pictureCaption: $pictureCaption)';
+  }
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is Captions &&
+          listEquals(textCaptions, other.textCaptions) &&
+          pictureCaption == other.pictureCaption;
+
+  @override
+  int get hashCode => Object.hash(textCaptions, pictureCaption);
 }