+++ packages/video_player_avplay/tizen/src/video_player.cc
@@ -149,15 +149,17 @@ void VideoPlayer::SendBufferingEnd() {
   PushEvent(flutter::EncodableValue(result));
 }
 
-void VideoPlayer::SendSubtitleUpdate(int32_t duration, const std::string &text,
-                                     flutter::EncodableList attributes) {
+void VideoPlayer::SendSubtitleUpdate(int32_t duration,
+                                     flutter::EncodableList texts_info,
+                                     flutter::EncodableMap picture_info) {
   flutter::EncodableMap result = {
       {flutter::EncodableValue("event"),
        flutter::EncodableValue("subtitleUpdate")},
       {flutter::EncodableValue("duration"), flutter::EncodableValue(duration)},
-      {flutter::EncodableValue("text"), flutter::EncodableValue(text)},
-      {flutter::EncodableValue("attributes"),
-       flutter::EncodableValue(attributes)},
+      {flutter::EncodableValue("textsInfo"),
+       flutter::EncodableValue(texts_info)},
+      {flutter::EncodableValue("pictureInfo"),
+       flutter::EncodableValue(picture_info)},
   };
 
   PushEvent(flutter::EncodableValue(result));