+++ packages/video_player_avplay/tizen/src/video_player.h
@@ -83,8 +83,8 @@ class VideoPlayer {
   void SendBufferingUpdate(int32_t value);
   void SendBufferingEnd();
   void SendSubtitleUpdate(
-      int32_t duration, const std::string &text,
-      flutter::EncodableList attributes = flutter::EncodableList{});
+      int32_t duration, flutter::EncodableList texts_info,
+      flutter::EncodableMap picture_info = flutter::EncodableMap{});
   void SendPlayCompleted();
   void SendIsPlayingState(bool is_playing);
   void SendRestored();